refactor: update auth styles and update directory
