+++ src/components/app/app.js
@@ -2,9 +2,11 @@ import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
 import Main from 'routes/main';
-import Login from 'routes/login';
-import Signup from 'routes/signup';
-import ForgotPassword from 'routes/forgot-password';
+import Login from 'routes/auth/login';
+import Signup from 'routes/auth/signup';
+import ForgotPassword from 'routes/auth/forgot-password';
+
+import 'routes/auth/styles.scss';
 
 const App = () => {
   return (
