+++ src/components/app/app.scss
@@ -1,12 +1,5 @@
 @charset "utf-8";
 
-.app {
-  text-align: center;
-}
-
-// Import a Google Font
-// @import url("https://fonts.googleapis.com/css?family=Nunito:400,700");
-
 // Set your brand colors
 // $purple: #8a4d76;
 // $pink: #fa7c91;
