+++ src/routes/auth/styles.scss
@@ -1,0 +1,27 @@
+.container-wrapper {
+  margin-top: 3rem;
+  background-color: white;
+  border-radius: 5px;
+  box-shadow: 0 2px 20px 0 #e1e0e9;
+  position: relative;
+  top: 5px;
+  padding: 1.5rem !important;
+
+  h1.form-title {
+    font-weight: bold;
+    font-size: 1.5rem;
+    text-align: center;
+    margin: 0 auto;
+    margin-bottom: 3rem;
+  }
+
+  .form-wrapper {
+    max-width: 24rem;
+    margin: auto;
+    margin-bottom: 2rem;
+  }
+}
+
+body {
+  background: linear-gradient(to bottom, #f6f6f9 0%, #ffffff 100%);
+}
