--- src/routes/login/Login.scss
@@ -1,8 +1,0 @@
-.field.is-grouped {
-  justify-content: space-between !important;
-  align-items: center;
-
-  .forget-password-link {
-    font-size: 0.95rem;
-  }
-}
